# server/openapi/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "DD Core API",
    "version": "0.1.0",
    "description": "Stable v1 contract for DD instance discovery, service health, account authentication, contacts, and P4 reliable text messaging/sync. P0 call endpoints remain experimental and are intentionally excluded until promoted into /api/v1.",
    "license": {
      "name": "License decision pending",
      "url": "https://dd.invalid/license-pending"
    }
  },
  "servers": [
    {
      "url": "https://{instanceHost}",
      "description": "Self-hosted DD instance",
      "variables": {
        "instanceHost": {
          "default": "dd.invalid",
          "description": "Hostname of the self-hosted instance"
        }
      }
    }
  ],
  "security": [],
...
  "paths": {
    "/.well-known/openimx/client": {
      "get": {
        "operationId": "getWellKnownClientConfiguration",
        "summary": "Discover a DD instance",
        "description": "Returns the minimal unauthenticated connection document. This endpoint is intentionally not wrapped in the normal API envelope.",
        "tags": ["Discovery"],
        "responses": {
          "200": {
            "description": "Client discovery document",
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/RequestId"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Instance"
                }
              }
            }
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          }
        }
      }
    },
    "/api/v1/instance": {
      "get": {
        "operationId": "getInstance",
        "summary": "Get instance configuration",
        "tags": ["Discovery"],
        "responses": {
          "200": {
            "description": "Instance configuration",
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/RequestId"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstanceResponse"
                }
              }
            }
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          }
        }
      }
    },
    "/api/v1/auth/register/email/send-code": {
      "post": {
        "operationId": "requestEmailCode",
        "summary": "Request an email verification code",
        "description": "Requests a short-lived registration code. Accepted responses do not reveal whether the email is already registered.",
        "tags": ["Auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {"$ref": "#/components/schemas/EmailCodeRequest"}
            }
          }
        },
        "responses": {
          "202": {
            "description": "Request accepted",
            "headers": {"X-Request-ID": {"$ref": "#/components/headers/RequestId"}, "Retry-After": {"schema": {"type": "integer", "minimum": 1}}},
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/EmailCodeAcceptedResponse"}}}
          },
          "400": {"description": "Invalid email or purpose", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}},
          "403": {"description": "Registration is disabled", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}},
          "429": {"description": "Verification-code rate limit exceeded", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}},
          "503": {"description": "Authentication service is unavailable", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}}
        }
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "operationId": "registerAccount",
        "summary": "Register a new account",
        "description": "Atomically consumes a registration code and creates the user, privacy settings, Argon2id password record, first device and refresh-token family.",
        "tags": ["Auth"],
        "requestBody": {
          "required": true,
          "content": {"application/json": {"schema": {"$ref": "#/components/schemas/RegisterRequest"}}}
        },
        "responses": {
          "201": {"description": "Account created and authenticated", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/AuthSessionResponse"}}}},
          "400": {"description": "Input or verification code is invalid", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}},
          "403": {"description": "Registration is disabled", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}},
          "409": {"description": "Email or handle is already in use", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}},
          "503": {"description": "Authentication service is unavailable", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}}
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "operationId": "loginWithPassword",
        "summary": "Login with email and password",
        "description": "Uses a unified invalid-credentials response for unknown accounts and wrong passwords, creates a device session and starts a refresh-token family.",
        "tags": ["Auth"],
        "requestBody": {
          "required": true,
          "content": {"application/json": {"schema": {"$ref": "#/components/schemas/LoginRequest"}}}
        },
        "responses": {
          "200": {"description": "Authenticated session", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/AuthSessionResponse"}}}},
          "400": {"description": "Request shape or device metadata is invalid", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}},
          "401": {"description": "Email or password is incorrect", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}},
          "429": {"description": "Too many failed login attempts", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}},
          "503": {"description": "Authentication service is unavailable", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}}
        }
      }
    },
    "/api/v1/auth/token/refresh": {
      "post": {
        "operationId": "rotateRefreshToken",
        "summary": "Rotate a refresh token",
        "description": "Consumes one refresh token and returns a new access/refresh pair. Reusing a consumed refresh token revokes its entire token family.",
        "tags": ["Auth"],
        "requestBody": {
          "required": true,
          "content": {"application/json": {"schema": {"$ref": "#/components/schemas/RefreshRequest"}}}
        },
        "responses": {
          "200": {"description": "Refresh token rotated", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/AuthSessionResponse"}}}},
          "400": {"description": "Request shape is invalid", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}},
          "401": {"description": "Refresh token is invalid, expired, revoked or reused", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}},
          "503": {"description": "Authentication service is unavailable", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}}
        }
      }
    },
    "/api/v1/auth/password/reset/send-code": {
      "post": {
        "operationId": "requestPasswordResetCode",
        "summary": "Request a password reset code",
        "description": "Always uses accepted semantics for syntactically valid requests so callers cannot enumerate registered emails.",
        "tags": ["Auth"],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PasswordResetCodeRequest"}}}},
        "responses": {
          "202": {"description": "Request accepted", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/EmailCodeAcceptedResponse"}}}},
          "400": {"description": "Request is malformed", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}},
          "429": {"description": "Reset-code rate limit exceeded", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}}
        }
      }
    },
    "/api/v1/auth/password/reset": {
      "post": {
        "operationId": "resetPassword",
        "summary": "Reset the account password",
        "description": "Consumes a short-lived one-time code, updates the Argon2id password hash and revokes every existing device and refresh-token session.",
        "tags": ["Auth"],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PasswordResetRequest"}}}},
        "responses": {
          "200": {"description": "Password reset and existing sessions revoked", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/OperationResultResponse"}}}},
          "400": {"description": "Code, password or request is invalid", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}}
        }
      }
    },
    "/api/v1/auth/logout-all": {
      "post": {
        "operationId": "logoutAllDevices",
        "summary": "Revoke every device session",
        "tags": ["Auth"],
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {"description": "All device sessions revoked", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/OperationResultResponse"}}}},
          "401": {"description": "Access token is missing or invalid", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}}
        }
      }
    },
    "/api/v1/me": {
      "get": {
        "operationId": "getCurrentUser",
        "summary": "Get current profile and privacy settings",
        "tags": ["Auth"],
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {"description": "Current profile", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/AccountMeResponse"}}}},
          "401": {"description": "Access token is missing or invalid", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}}
        }
      },
      "patch": {
        "operationId": "updateCurrentUser",
        "summary": "Update profile and privacy settings",
        "tags": ["Auth"],
        "security": [{"bearerAuth": []}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UpdateMeRequest"}}}},
        "responses": {
          "200": {"description": "Updated profile", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/AccountMeResponse"}}}},
          "400": {"description": "Profile input is invalid", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}},
          "401": {"description": "Access token is missing or invalid", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}}
        }
      }
    },
    "/api/v1/me/email/send-code": {
      "post": {
        "operationId": "requestCurrentUserEmailChangeCode",
        "summary": "Send a verification code for changing the current user's email",
        "tags": ["Auth"],
        "security": [{"bearerAuth": []}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/EmailChangeCodeRequest"}}}},
        "responses": {
          "202": {"description": "Email change verification request accepted", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/EmailCodeAcceptedResponse"}}}},
          "400": {"description": "Email is invalid or matches the current email", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}},
          "401": {"description": "Access token is missing or invalid", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}},
          "409": {"description": "Email is already registered", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}},
          "503": {"description": "Verification delivery is temporarily unavailable", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}}
        }
      }
    },
...
    "/api/v1/devices": {
      "get": {
        "operationId": "listCurrentUserDevices",
        "summary": "List the current user's devices",
        "tags": ["Auth"],
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {"description": "Device list", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/DeviceListResponse"}}}},
          "401": {"description": "Access token is missing or invalid", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}}
        }
      }
    },
...
  },
  "components": {
...
    "headers": {
      "RequestId": {
        "description": "Opaque identifier used to correlate one HTTP request across logs.",
        "schema": {
          "type": "string",
          "pattern": "^req_[A-Za-z0-9]+$"
        }
      }
    },
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Instance": {
        "type": "object",
        "additionalProperties": false,
        "required": ["name", "apiVersion", "apiBaseUrl", "realtimeUrl", "liveKitUrl", "features"],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 80
          },
          "apiVersion": {
            "type": "string",
            "const": "v1"
          },
          "apiBaseUrl": {
            "type": "string",
            "format": "uri"
          },
          "realtimeUrl": {
            "type": "string",
            "format": "uri"
          },
          "liveKitUrl": {
            "type": "string",
            "format": "uri"
          },
          "features": {
            "$ref": "#/components/schemas/FeatureFlags"
          }
        }
      },
...
    }
  }
}